Remove ridf.file & Implement parser
